LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.arc.nonfinite-expected.txt

2d.path.arc.nonfinite

arc() with Infinity/NaN is ignored


PASS arc() with Infinity/NaN is ignored 
